Switching approach in building Sliders component.
Tried to build Sliders using a composition of Spark's Slider and elements of its inheritance
chain, and though I made it decently far, in the end it proved to be too much of trying to force a square
peg in a round hole.

Going to try to build a Slider component from scratch instead using a layout approach I believe
the Flex team ought to have used instead of the long inheritance chains and behemoth classes.