Made Thumb focus and basic movement via keys work.
Thumb doesnt move if snapInterval is 0. Need to address this somehow, either using
"stepSize" or defining a specific movement interval when snapInterval is 0.

Had to make SliderThumb a SkinnableComp as opposed to a SkinnableCont in order
to make focus work. This in turn required tests using a SliderThumb to define
a skin for it. If possible, it needs to be mocked.